requirements.txt
python-dotenv
google-generativeai
structlog
backoff